Tidy up demo pages
