+++ packages/scxa-faceted-search-results/html/faceted-search-container-demo.html
@@ -30,7 +30,6 @@
 </body>
 
 <script src="../dist/vendors.bundle.js"></script>
-<script src="../dist/facetedSearch.bundle.js"></script>
 <script src="../dist/facetedSearchContainerDemo.bundle.js"></script>
 
 <!-- Set to http://localhost:8080/gxa/ or http://localhost:8080/gxa_sc/ -- Remember the trailing slash! -->
