+++ packages/scxa-faceted-search-results/html/fetch-loader/demo.html
@@ -25,7 +25,6 @@
 </body>
 
 <script src="../../dist/vendors.bundle.js"></script>
-<script src="../../dist/facetedSearch.bundle.js"></script>
 <script src="../../dist/fetchLoaderDemo.bundle.js"></script>
 
 <!-- Set to http://localhost:8080/gxa/ or http://localhost:8080/gxa/sc/ -- Remember the trailing slash! -->
@@ -33,7 +32,7 @@
     const symbol = `Aspa`;
 
     fetchLoaderDemo.render({
-      host: 'http://localhost:8080/gxa/sc/',
+      host: 'https://www-test.ebi.ac.uk/gxa/sc/',
       resource: 'json/search?symbol=' + symbol, // Add species to see shared facets go away: + '&species=mus musculus',
       // matchingGeneId will be populated for species-specific genes
       resultsMessageFormatter: function(data) { return symbol + ' ' + data.matchingGeneId + ' is expressed in:'; },
