+++ packages/scxa-faceted-search-results/html/fetch-loader/error.html
@@ -17,7 +17,6 @@
 </body>
 
 <script src="../../dist/vendors.bundle.js"></script>
-<script src="../../dist/facetedSearch.bundle.js"></script>
 <script src="../../dist/fetchLoaderDemo.bundle.js"></script>
 
 <!-- Set to http://localhost:8080/gxa/ or http://localhost:8080/gxa/sc/ -- Remember the trailing slash! -->
